mas funciones

+++ src/Campaigns.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useMemo } from 'react';
+import React, { useState, useEffect, useMemo, useContext } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { Container, Row, Col, Button, Table, Form, InputGroup, FormControl, DropdownButton, Dropdown } from 'react-bootstrap';
 import TemplatePreview from './TemplatePreview';
@@ -7,16 +7,61 @@ import { io } from 'socket.io-client';
 import './Campaigns.css';
 import { ArrowUpSquare, CheckCircle, Clock, RocketFill, ThreeDotsVertical, XCircle } from 'react-bootstrap-icons';
 import Swal from 'sweetalert2';
+import { AppContext } from './context';
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+  Card,
+  CardContent,
+  CardDescription,
+  CardFooter,
+  CardHeader,
+  CardTitle,
+  ChartContainer,
+  ChartTooltip,
+  ChartTooltipContent,
+} from "./components"
+
+import {
+  Area,
+  AreaChart,
+  Bar,
+  BarChart,
+  CartesianGrid,
+  Label,
+  LabelList,
+  Line,
+  LineChart,
+  PolarAngleAxis,
+  RadialBar,
+  RadialBarChart,
+  Rectangle,
+  ReferenceLine,
+  XAxis,
+  YAxis,
+} from "recharts"
 
 const socket = io(process.env.REACT_APP_API_URL);
 
-const Campaigns = () => {
+export const Campaigns = () => {
+
+  const {setCampaigns: addCampaigns, setTemplates: addTemplates} = useContext(AppContext);
+
   const navigate = useNavigate();
   const [templates, setTemplates] = useState([]);
   const [searchTerm, setSearchTerm] = useState('');
   const [filterType, setFilterType] = useState('');
   const [selectedTemplate, setSelectedTemplate] = useState(null);
   const [campaigns, setCampaigns] = useState([]);
+  const [showDialog, setShowDialog] = useState(false);
+  const [selectedCampaign, setSelectedCampaign] = useState(null);
 
   useEffect(() => {
     const fetchTemplates = async () => {
@@ -28,13 +73,14 @@ const Campaigns = () => {
       }
 
       try {
-        console.log('Fetching templates with company ID:', companyId);
+      
         const response = await axios.get(`${process.env.REACT_APP_API_URL}/api/templates`, {
           params: { company_id: companyId },
           headers: { Authorization: `Bearer ${token}` }
         });
-        console.log('Fetched templates:', response.data);
+
         setTemplates(response.data);
+        addTemplates(response.data)
       } catch (error) {
         console.error('Error fetching templates:', error);
       }
@@ -51,6 +97,7 @@ const Campaigns = () => {
         });
         console.log('Fetched campaigns:', response.data);
         setCampaigns(response.data);
+        addCampaigns(response.data)
       } catch (error) {
         console.error('Error fetching campaigns:', error);
       }
@@ -111,7 +158,7 @@ const Campaigns = () => {
         } catch (error) {
           Swal.fire({
             title: "Error",
-            text: `Error al eliminar Liquidación.
+            text: `Error al eliminar Plantilla.
             Error: ${error}`,
             icon: "error"
           });
@@ -124,18 +171,49 @@ const Campaigns = () => {
     navigate(`/edit-campaign/${campaign.id}`);
   };
 
+  const handleDetailsCampaignClick = (campaign) => {
+    setSelectedCampaign(campaign); 
+    console.log(campaign)
+    setShowDialog(true); // Muestra el diálogo
+  };
+
+  const handleCloseDialog = () => {
+    setShowDialog(false); 
+    setSelectedCampaign(null); 
+  };
+
   const handleDeleteCampaignClick = async (campaignId) => {
     const token = localStorage.getItem('token');
-    try {
-      console.log('Deleting campaign with ID:', campaignId);
-      await axios.delete(`${process.env.REACT_APP_API_URL}/api/campaigns/${campaignId}`, {
-        headers: { Authorization: `Bearer ${token}` }
-      });
-      setCampaigns(campaigns.filter(campaign => campaign.id !== campaignId));
-    } catch (error) {
-      console.error('Error deleting campaign:', error);
-    }
-  };
+    Swal.fire({
+      title: "Esta seguro que desea eliminar esta Campaña?",
+      showDenyButton: true,
+      confirmButtonText: "Eliminar",
+    }).then(async (result) => { 
+      if (result.isConfirmed) {
+        try {
+          console.log('Deleting campaign with ID:', campaignId);
+          await axios.delete(`${process.env.REACT_APP_API_URL}/api/campaigns/${campaignId}`, {
+            headers: { Authorization: `Bearer ${token}` }
+          });
+          setCampaigns(campaigns.filter(campaign => campaign.id !== campaignId));
+          await Swal.fire({
+            title: "Perfecto",
+            text: `Campaña Eliminada.`,
+            icon: "success"
+          });
+        } catch (error) {
+          console.error('Error deleting campaign:', error);
+          Swal.fire({
+            title: "Error",
+            text: `Error al eliminar Campaña.
+            Error: ${error}`,
+            icon: "error"
+          });
+        }
+      }
+   
+    });
+  }
 
   const handleLaunchCampaignClick = async (campaignId) => {
     const token = localStorage.getItem('token');
@@ -278,7 +356,7 @@ const Campaigns = () => {
                           <RocketFill /> Lanzar
                         </Button>
                         <DropdownButton id="dropdown-basic-button" className="custom-dropdown-toggle" title={<ThreeDotsVertical />} variant="ghost" size="sm">
-                          <Dropdown.Item onClick={() => handleEditCampaignClick(campaign)}>
+                          <Dropdown.Item onClick={() => handleDetailsCampaignClick(campaign)}>
                             Detalles
                           </Dropdown.Item>
                           <Dropdown.Item onClick={() => handleEditCampaignClick(campaign)}>
@@ -305,8 +383,124 @@ const Campaigns = () => {
           </Col>
         </Row>
       </Col>
+      {showDialog && (
+      <AlertDialog open={showDialog} onOpenChange={setShowDialog}>
+        <AlertDialogTrigger asChild>
+          {/* Un solo botón que actúa como trigger */}
+          <button style={{ display: 'none' }}>Open</button>
+        </AlertDialogTrigger>
+        <AlertDialogContent>
+          <AlertDialogHeader>
+            <AlertDialogTitle>Detalles de la campaña</AlertDialogTitle>
+            <AlertDialogDescription>
+              <Card
+                  className="max-w-full" x-chunk="charts-01-chunk-5"
+                >
+                  <CardContent className="flex gap-4 p-4">
+                    <div className="grid items-center gap-2">
+                      <div className="grid flex-1 auto-rows-min gap-0.5">
+                        <div className="text-sm text-muted-foreground">Total enviadas</div>
+                        <div className="flex items-baseline gap-1 text-xl font-bold tabular-nums leading-none">
+                          {`${selectedCampaign.interactions}`}/{`${selectedCampaign.conversions}`}
+                          <span className="text-sm font-normal text-muted-foreground">
+                            Personas
+                          </span>
+                        </div>
+                      </div>
+                      <div className="grid flex-1 auto-rows-min gap-0.5">
+                        <div className="text-sm text-muted-foreground">Recibidas</div>
+                        <div className="flex items-baseline gap-1 text-xl font-bold tabular-nums leading-none">
+                        {`${selectedCampaign.delivered}`}
+                          <span className="text-sm font-normal text-muted-foreground">
+                            Mensajes
+                          </span>
+                        </div>
+                      </div>
+                      <div className="grid flex-1 auto-rows-min gap-0.5">
+                        <div className="text-sm text-muted-foreground">Leídas</div>
+                        <div className="flex items-baseline gap-1 text-xl font-bold tabular-nums leading-none">
+                        {`${selectedCampaign.read}`}
+                          <span className="text-sm font-normal text-muted-foreground">
+                            Mensajes
+                          </span>
+                        </div>
+                      </div>
+                    </div>
+                    <ChartContainer
+                      config={{
+                        move: {
+                          label: "Totales",
+                          color: "hsl(var(--chart-1))",
+                        },
+                        exercise: {
+                          label: "Recibidas",
+                          color: "hsl(var(--chart-2))",
+                        },
+                        stand: {
+                          label: "Leídas",
+                          color: "hsl(var(--chart-3))",
+                        },
+                      }}
+                      className="mx-auto aspect-square w-full max-w-[80%]"
+                    >
+                      <RadialBarChart
+                        margin={{
+                          left: -10,
+                          right: -10,
+                          top: -10,
+                          bottom: -10,
+                        }}
+                        data={[
+                          {
+                            name: "Leidos",
+                            activity: "stand",
+                            value: (selectedCampaign.delivered / selectedCampaign.interactions)* 100,
+                            fill: "var(--color-stand)",
+                          },
+                          {
+                            name: "Recibidos",
+                            activity: "exercise",
+                            value: (selectedCampaign.read / selectedCampaign.interactions) * 100,
+                            fill: "var(--color-exercise)",
+                          },
+                          {
+                            name: "Enviados",
+                            activity: "move",
+                            value: (selectedCampaign.interactions / selectedCampaign.conversions) * 100,
+                            fill: "var(--color-move)",
+                          },
+                        ]}
+                        innerRadius="20%"
+                        barSize={27}
+                        startAngle={90}
+                        endAngle={450}
+                      >
+                        <PolarAngleAxis
+                          type="number"
+                          domain={[0, 100]}
+                          dataKey="value"
+                          tick={false}
+                        />
+                        <RadialBar dataKey="value" background cornerRadius={5} > 
+                          <LabelList
+                            position="insideStart"
+                            dataKey="name"
+                            className="fill-white capitalize mix-blend-luminosity"
+                            fontSize={15}
+                          />
+                        </RadialBar>
+                      </RadialBarChart>
+                    </ChartContainer>
+                  </CardContent>
+                </Card>
+            </AlertDialogDescription>
+          </AlertDialogHeader>
+          <AlertDialogFooter>
+            <AlertDialogCancel onClick={handleCloseDialog}>Cerrar</AlertDialogCancel>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
+     )}
     </Container>
   );
 };
-
-export default Campaigns;
